requirements.txt
fastapi
h11==0.14.0
httptools==0.5.0
motor==2.5.1
pymongo==3.12.0
pydantic
python-dotenv==1.0.0
starlette
typing-extensions==4.5.0
uvicorn==0.22.0
email-validator
